+++ src/main/java/codesquad/fineants/spring/api/portfolio_stock/event/PortfolioEvent.java
@@ -1,4 +1,4 @@
-package codesquad.fineants.spring.api.portfolio_stock;
+package codesquad.fineants.spring.api.portfolio_stock.event;
 
 import java.time.LocalDateTime;
