+++ src/main/java/codesquad/fineants/spring/api/portfolio_stock/manager/SseEmitterManager.java
@@ -1,4 +1,4 @@
-package codesquad.fineants.spring.api.portfolio_stock;
+package codesquad.fineants.spring.api.portfolio_stock.manager;
 
 import java.util.Map;
 import java.util.Set;
