+++ src/main/java/codesquad/fineants/spring/api/portfolio_stock/service/StockMarketChecker.java
@@ -1,4 +1,4 @@
-package codesquad.fineants.spring.api.portfolio_stock;
+package codesquad.fineants.spring.api.portfolio_stock.service;
 
 import java.time.DayOfWeek;
 import java.time.LocalDateTime;
