+++ src/test/java/codesquad/fineants/spring/api/portfolio_stock/PortfolioEventPublisherTest.java
@@ -36,6 +36,9 @@ import codesquad.fineants.domain.stock_dividend.StockDividendRepository;
 import codesquad.fineants.spring.api.kis.manager.CurrentPriceManager;
 import codesquad.fineants.spring.api.kis.manager.LastDayClosingPriceManager;
 import codesquad.fineants.spring.api.kis.response.CurrentPriceResponse;
+import codesquad.fineants.spring.api.portfolio_stock.event.PortfolioEventPublisher;
+import codesquad.fineants.spring.api.portfolio_stock.manager.SseEmitterManager;
+import codesquad.fineants.spring.api.portfolio_stock.service.PortfolioStockService;
 
 @ActiveProfiles("test")
 @SpringBootTest
